dots/requirements.txt
customtkinter
Pillow
streamlit
